+++ module/opcode.cpp
@@ -3,48 +3,48 @@
 #include "module/video.hpp"
 
 namespace op {
- u8 push(u8 value) {
+ u32 push(u32 value) {
   if (stacker > 0) {memory[--stacker] = value;}
   return 0;
  }
 
- u8 pop(u8 value) {
+ u32 pop(u32 value) {
   if (stacker < SYSTEM::MEMORY) {return memory[stacker++];}
   return 0;
  }
 
- u8 pushm(u8 value) {
+ u32 pushm(u32 value) {
   if (stacker > 0) {memory[--stacker] = memory[value];}
   return 0;
  }
 
- u8 popm(u8 value) {
+ u32 popm(u32 value) {
   if (stacker < SYSTEM::MEMORY) {
    memory[value] = memory[stacker++];
   }
   return 0;
  }
 
- u8 clear(u8 value) {
-  u8 color = pop(0);
+ u32 clear(u32 value) {
+  u32 color = pop(0);
   video::clear(color);
   return 0;
  }
 
- u8 pixel(u8 value) {
-  u8 color = pop(0);
-  u8 y = pop(0);
-  u8 x = pop(0);
+ u32 pixel(u32 value) {
+  u32 color = pop(0);
+  u32 y = pop(0);
+  u32 x = pop(0);
   video::pixel(x, y, color);
   return 0;
  }
 
- u8 flip(u8 value) {
+ u32 flip(u32 value) {
   video::flip();
   return 0;
  }
 
- u8 nop(u8 value) {
+ u32 nop(u32 value) {
   return 0;
  }
 }
